# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSchemaModelBindingRequest": {
        "properties": {
          "modelId": {
            "format": "int64",
            "type": "integer"
          },
          "pluginPolicy": {
            "additionalProperties": { },
            "type": [
              "object",
              "null"
            ]
          }
        },
        "required": [
          "modelId"
        ],
        "type": "object"
      },
      "CreateSchemaRequest": {
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "CreateSchemaVersionRequest": {
        "properties": {
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/CreateSchemaModelBindingRequest"
            },
            "minItems": 1,
            "type": "array"
          },
          "formSchema": {
            "additionalProperties": { },
            "minProperties": 1,
            "type": "object"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "bindings",
          "formSchema"
        ],
        "type": "object"
      },
...
      "OrganizationDto": {
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "avatarUrl",
          "createdAt",
          "description",
          "id",
          "name",
          "slug",
          "updatedAt"
        ],
        "type": "object"
      },
      "OrganizationMembershipDto": {
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "roleDefinition": {
            "$ref": "#/components/schemas/RoleSummaryDto"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "PENDING",
              "REMOVED"
            ],
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "avatarUrl",
          "createdAt",
          "email",
          "fullName",
          "id",
          "organizationId",
          "roleDefinition",
          "status",
          "userId"
        ],
        "type": "object"
      },
...
      "PageDtoSchemaCatalogItemDto": {
        "properties": {
          "hasNext": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SchemaCatalogItemDto"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalItems": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "hasNext",
          "items",
          "page",
          "size",
          "totalItems"
        ],
        "type": "object"
      },
...
      "RoleSummaryDto": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "locked": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "scope": {
            "enum": [
              "SYSTEM",
              "ORGANIZATION"
            ],
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "systemKey": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "locked",
          "name",
          "scope",
          "slug",
          "systemKey"
        ],
        "type": "object"
      },
...
      "SchemaCatalogItemDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fieldCount": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "modelCount": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "reportCount": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedByAvatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedByEmail": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedByName": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "archivedAt",
          "createdAt",
          "description",
          "fieldCount",
          "id",
          "modelCount",
          "name",
          "organizationId",
          "reportCount",
          "updatedAt",
          "updatedByAvatarUrl",
          "updatedByEmail",
          "updatedByName"
        ],
        "type": "object"
      },
...
      "SchemaDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "archivedAt",
          "createdAt",
          "description",
          "id",
          "name",
          "organizationId",
          "updatedAt"
        ],
        "type": "object"
      },
      "SchemaModelBindingDto": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "modelId": {
            "format": "int64",
            "type": "integer"
          },
          "modelName": {
            "type": "string"
          },
          "pluginPolicy": {
            "additionalProperties": { },
            "type": [
              "object",
              "null"
            ]
          },
          "schemaVersionId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "modelId",
          "modelName",
          "pluginPolicy",
          "schemaVersionId"
        ],
        "type": "object"
      },
...
      "SchemaVersionDto": {
        "properties": {
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/SchemaModelBindingDto"
            },
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "formSchema": {
            "additionalProperties": { },
            "type": "object"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "schemaId": {
            "format": "int64",
            "type": "integer"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "bindings",
          "createdAt",
          "formSchema",
          "id",
          "name",
          "schemaId",
          "version"
        ],
        "type": "object"
      },
...
      "SelectOrganizationRequest": {
        "properties": {
          "organizationId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "organizationId"
        ],
        "type": "object"
      },
...
      "WorkspaceContextDto": {
        "properties": {
          "currentMembership": {
            "$ref": "#/components/schemas/OrganizationMembershipDto"
          },
          "currentOrganization": {
            "$ref": "#/components/schemas/OrganizationDto"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/OrganizationMembershipDto"
            },
            "type": "array"
          },
          "organizations": {
            "items": {
              "$ref": "#/components/schemas/OrganizationDto"
            },
            "type": "array"
          },
          "permissions": {
            "$ref": "#/components/schemas/WorkspacePermissionsDto"
          },
          "user": {
            "$ref": "#/components/schemas/WorkspaceUserDto"
          }
        },
        "required": [
          "currentMembership",
          "currentOrganization",
          "memberships",
          "organizations",
          "permissions",
          "user"
        ],
        "type": "object"
      },
      "WorkspacePermissionsDto": {
        "properties": {
          "canCreateModels": {
            "type": "boolean"
          },
          "canDeleteModels": {
            "type": "boolean"
          },
          "canDeleteOrganization": {
            "type": "boolean"
          },
          "canEditModels": {
            "type": "boolean"
          },
          "canEditOrganization": {
            "type": "boolean"
          },
          "canExportPredictions": {
            "type": "boolean"
          },
          "canInviteMembers": {
            "type": "boolean"
          },
          "canManageInvitations": {
            "type": "boolean"
          },
          "canManageMemberRoles": {
            "type": "boolean"
          },
          "canManagePlugins": {
            "type": "boolean"
          },
          "canManageReviews": {
            "type": "boolean"
          },
          "canRemoveMembers": {
            "type": "boolean"
          },
          "canReview": {
            "type": "boolean"
          },
          "canRunPredictions": {
            "type": "boolean"
          },
          "canTransferOwnership": {
            "type": "boolean"
          },
          "canViewInvitations": {
            "type": "boolean"
          },
          "canViewMembers": {
            "type": "boolean"
          },
          "canViewModels": {
            "type": "boolean"
          },
          "canViewOrganization": {
            "type": "boolean"
          },
          "canViewPlugins": {
            "type": "boolean"
          },
          "canViewWorkspace": {
            "type": "boolean"
          }
        },
        "required": [
          "canCreateModels",
          "canDeleteModels",
          "canDeleteOrganization",
          "canEditModels",
          "canEditOrganization",
          "canExportPredictions",
          "canInviteMembers",
          "canManageInvitations",
          "canManageMemberRoles",
          "canManagePlugins",
          "canManageReviews",
          "canRemoveMembers",
          "canReview",
          "canRunPredictions",
          "canTransferOwnership",
          "canViewInvitations",
          "canViewMembers",
          "canViewModels",
          "canViewOrganization",
          "canViewPlugins",
          "canViewWorkspace"
        ],
        "type": "object"
      },
      "WorkspaceUserDto": {
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "avatarUrl",
          "email",
          "fullName",
          "id"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MLSuite API",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analyzer/schema": {
      "post": {
        "operationId": "generateSchema",
        "parameters": [
          {
            "in": "query",
            "name": "oneHotSeparator",
            "required": false,
            "schema": {
              "default": "__",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "dataframe": {
                    "format": "binary",
                    "type": "string"
                  },
                  "model": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "model"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": { },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "analyzer-controller"
        ]
      }
    },
...
    "/api/schema-versions/{versionId}": {
      "get": {
        "operationId": "get_3",
        "parameters": [
          {
            "in": "path",
            "name": "versionId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaVersionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-version-controller"
        ]
      }
    },
    "/api/schema-versions/{versionId}/bindings": {
      "get": {
        "operationId": "listBindings",
        "parameters": [
          {
            "in": "path",
            "name": "versionId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SchemaModelBindingDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-version-controller"
        ]
      },
      "post": {
        "operationId": "addBinding",
        "parameters": [
          {
            "in": "path",
            "name": "versionId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaModelBindingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaModelBindingDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-version-controller"
        ]
      }
    },
    "/api/schemas": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 24,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDtoSchemaCatalogItemDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      }
    },
...
    "/api/schemas/{schemaId}": {
      "delete": {
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      },
      "get": {
        "operationId": "get_2",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      },
      "patch": {
        "operationId": "rename",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      }
    },
...
    "/api/schemas/{schemaId}/versions": {
      "get": {
        "operationId": "list_1",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SchemaVersionDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-version-controller"
        ]
      },
      "post": {
        "operationId": "create_1",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaVersionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-version-controller"
        ]
      }
    },
...
    "/api/workspace/context": {
      "get": {
        "operationId": "getContext",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceContextDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "workspace-controller"
        ]
      },
      "patch": {
        "operationId": "selectOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceContextDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "workspace-controller"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}